YTDB-645: merge GapCoverage into SchemaVariations, fix review findings
